Add user, document, review, comment, insertion and deletion scaffolds
Create scaffolds to represent a user, a document, a review, a comment,
an insertion and a deletion. A user has a name and an email that for
now can be anything.

A document belongs to a user and has a name, some content and a
revision number. Every time a review is accepted the revision number
is incremented.

A review belongs to a user and to a document. Reviews contains several
comments, insertions and deletions. A review also has a base revision
of the document to which these changes should be applied.

A comment belongs to a review. It has the position of the comment in
the text, the offset in the text over which it is applied and some
content.

An insertion belongs to a review. It has the position of the insertion
in the text and the content to insert.

A deletion belongs to a review. It has the position of the deletion in
the text and the offset over which to delete.